# ecommerce/docs/swagger.yaml
definitions:
...
  controllers.reviewProductInput:
    properties:
      deskripsi:
        type: string
      product_id:
        type: integer
      rating:
        type: integer
      user_id:
        type: integer
    type: object
  controllers.saleOrderInput:
    properties:
      payment_id:
        type: integer
      shipping_address:
        type: string
      status:
        type: string
      users_id:
        type: integer
    type: object
...
  models.ReviewProduct:
    properties:
      created_at:
        type: string
      deskripsi:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      users_id:
        description: FK
        type: integer
    type: object
  models.SaleOrder:
    properties:
      created_at:
        type: string
      id:
        type: integer
      kode_pesanan:
        type: string
      order_date:
        type: string
      payment_id:
        type: integer
      shipping_address:
        type: string
      status:
        type: string
      updated_at:
        type: string
      users_id:
        description: FK
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /product/{id}/review-product:
    get:
      description: Get all ReviewProduct by product id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReviewProduct'
            type: array
      summary: Get a list of ReviewProducts.
      tags:
      - Review Product
...
  /review-product:
    get:
      description: Get a list of ReviewProduct.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReviewProduct'
            type: array
      summary: Get all Review Products.
      tags:
      - Review Product
    post:
      description: Creating a new ReviewProducts.
      parameters:
      - description: the body to create a new ReviewProduct
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.reviewProductInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReviewProduct'
      security:
      - BearerToken: []
      summary: Create New Review Products.
      tags:
      - Review Product
  /review-product/{id}:
    delete:
      description: Delete a ReviewProduct by id.
      parameters:
      - description: reviewProduct id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one ReviewProduct.
      tags:
      - Review Product
    put:
      description: Update ReviewProduct by id.
      parameters:
      - description: Review Product id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Review Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.reviewProductInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReviewProduct'
      security:
      - BearerToken: []
      summary: Update ReviewProducts.
      tags:
      - Review Product
  /sale-order:
    get:
      description: Get a list of saleorder.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SaleOrder'
            type: array
      summary: Get all Sale Orders.
      tags:
      - Sale Order
    post:
      description: Creating a new Sale orders.
      parameters:
      - description: the body to create a new SaleOrder
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.saleOrderInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaleOrder'
      security:
      - BearerToken: []
      summary: Create New Sale Orders.
      tags:
      - Sale Order
...
  /sale-order/{id}:
    delete:
      description: Delete a SO by id.
      parameters:
      - description: so id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Sale Orders.
      tags:
      - Sale Order
    put:
      description: Update Sale Order by id.
      parameters:
      - description: SaleOrder id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Sale Order
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.saleOrderInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaleOrder'
      security:
      - BearerToken: []
      summary: Update Sale orders.
      tags:
      - Sale Order
...
  /users/{id}/review-product:
    get:
      description: Get all ReviewProduct by user id
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReviewProduct'
            type: array
      summary: Get a list of ReviewProducts.
      tags:
      - User
swagger: "2.0"